usando sharedpreferences pra versionamento, implementando eventBus, integrando tudo com firebase

+++ app/src/main/java/br/com/manygames/agenda/dao/AlunoDAO.java
@@ -6,6 +6,7 @@ import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 import android.support.annotation.NonNull;
+import android.util.Log;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -16,7 +17,7 @@ import br.com.manygames.agenda.modelo.Aluno;
 public class AlunoDAO extends SQLiteOpenHelper{
 
     public AlunoDAO(Context context) {
-        super(context, "Agenda", null, 4);
+        super(context, "Agenda", null, 6);
     }
 
     @Override
@@ -28,7 +29,9 @@ public class AlunoDAO extends SQLiteOpenHelper{
                 "telefone TEXT, " +
                 "site TEXT, " +
                 "nota REAL, " +
-                "caminhoFoto TEXT)";
+                "caminhoFoto TEXT," +
+                "sincronizado INT DEFAULT 0," +
+                "desativado INT DEFAULT 0)";
         db.execSQL(sql);
     }
 
@@ -70,8 +73,12 @@ public class AlunoDAO extends SQLiteOpenHelper{
                      alunos) {
                     db.execSQL(atualizaIdAlunos, new String[]{geraUUID(), aluno.getId()});
                 }
-//            case 4:
-//                corrigeIdsNulos(db, "UPDATE Alunos SET id = ? where id IS NULL");
+            case 4:
+                String adicionaCampoSincroniado = "ALTER TABLE Alunos ADD COLUMN sincronizado INT DEFAULT 0";
+                db.execSQL(adicionaCampoSincroniado);
+            case 5:
+                String adicionaCampoDesativado = "ALTER TABLE Alunos ADD COLUMN desativado INT DEFAULT 0";
+                db.execSQL(adicionaCampoDesativado);
         }
     }
 
@@ -91,7 +98,7 @@ public class AlunoDAO extends SQLiteOpenHelper{
     }
 
     public List<Aluno> buscaAlunos() {
-        String sql = "SELECT * FROM Alunos;";
+        String sql = "SELECT * FROM Alunos WHERE desativado = 0";
         SQLiteDatabase db = getReadableDatabase();
         Cursor c = db.rawQuery(sql, null);
 
@@ -112,6 +119,8 @@ public class AlunoDAO extends SQLiteOpenHelper{
             aluno.setSite(c.getString(c.getColumnIndex("site")));
             aluno.setNota(c.getDouble(c.getColumnIndex("nota")));
             aluno.setCaminhoFoto(c.getString(c.getColumnIndex("caminhoFoto")));
+            aluno.setSincronizado(c.getInt(c.getColumnIndex("sincronizado")));
+            aluno.setDesativado(c.getInt(c.getColumnIndex("desativado")));
 
             alunos.add(aluno);
         }
@@ -121,7 +130,14 @@ public class AlunoDAO extends SQLiteOpenHelper{
     public void deleta(Aluno aluno) {
         SQLiteDatabase db = getWritableDatabase();
         String[] params =  {aluno.getId().toString()};
-        db.delete("Alunos", "id = ?", params);
+
+        if(aluno.estaDesativado()){
+            db.delete("Alunos", "id = ?", params);
+            Log.i("DELETADO!","PRA SEMPRE!");
+        } else {
+            aluno.desativa();
+            altera(aluno);
+        }
     }
 
     public void altera(Aluno aluno) {
@@ -156,6 +172,8 @@ public class AlunoDAO extends SQLiteOpenHelper{
         dados.put("site", aluno.getSite());
         dados.put("nota", aluno.getNota());
         dados.put("caminhoFoto", aluno.getCaminhoFoto());
+        dados.put("sincronizado", aluno.getSincronizado());
+        dados.put("desativado", aluno.getDesativado());
         return dados;
     }
 
@@ -170,6 +188,9 @@ public class AlunoDAO extends SQLiteOpenHelper{
     public void sincroniza(List<Aluno> alunos) {
         for (Aluno aluno:
              alunos) {
+
+            aluno.sincroniza();
+
             if(existe(aluno))
                 if(aluno.estaDesativado()){
                     deleta(aluno);
@@ -187,5 +208,13 @@ public class AlunoDAO extends SQLiteOpenHelper{
         String existe = "SELECT id FROM Alunos where id = ? LIMIT 1";
         Cursor c = db.rawQuery(existe, new String[]{aluno.getId()});
         return c.getCount() > 0;
+    }
+
+    public List<Aluno> listaNaoSincronizados(){
+        SQLiteDatabase db = getWritableDatabase();
+        String sincronizados = "SELECT * FROM Alunos WHERE sincronizado = 0";
+        Cursor c = db.rawQuery(sincronizados, null);
+        List<Aluno> alunos = populaAlunos(c);
+        return alunos;
     }
 }
